Give each index a unique ID
The ID is a string that can be obtained via the API with
cmd=get-index-id. It is currently just a concatenation of the index name
and several statistics on the number of triples (which is not perfect
because an index that is different but has the same such statistics will
get the same ID).

This feature is useful for applications like
https://github.com/ad-freiburg/qlever-petrimaps, which have an internal
cache that depends on the index. When the index changes, the application
should be able to notice this an clear its cache. With an index ID it
can.