Enable transitive paths of length zero.
Prior to this change, an exception was raised if a property path may
have length zero. Also, the methods in TransitivePath did not account
for the corner cases of the empty path. Especially the copyColumns
function goes out of bounds in a lot of cases and it was completely
refactored.

This change disables the exceptions for empty paths and instead
introduces this feature to the TransitivePath class. Computing all
results with an empty path can be very expensive, so the class will try
to choose the least amount of nodes to compute the transitive hull.
This choice happens in the function TransitivePath::setupMapAndNodes.
Only if both sides are unbound, the full transitive hull with the empty
paths of all nodes is computed.

Type: feature