Prepare the QueryPlanner for seedFrom... removal
Prior to this change, the logic of the seedFromPropertyPath function was
split over multiple functions. It also used a lot of shared_ptrs, which
are unnecessary in this case.

This change moves functions such as `seedFromAlternative` to the
`seedFromPropertyPath` function. The logic for alternatives, sequences
and so on will be moved in a future commit. Having it in a single place
instead of multiple functions makes this easier.